+++ src/main/java/com/ss/lms/LmsApplication.java
@@ -1,14 +1,20 @@
 package com.ss.lms;
 
+import javax.sql.DataSource;
+
 import org.apache.catalina.Context;
 import org.apache.catalina.connector.Connector;
 import org.apache.tomcat.util.descriptor.web.SecurityCollection;
 import org.apache.tomcat.util.descriptor.web.SecurityConstraint;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
 import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
 import org.springframework.context.annotation.Bean;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.provisioning.JdbcUserDetailsManager;
 
 @SpringBootApplication
 public class LmsApplication
@@ -38,6 +44,21 @@ public class LmsApplication
 	    tomcatServletWebServerFactory.addAdditionalTomcatConnectors(httpToHttpsRedirectConnector());
 	    
 	    return tomcatServletWebServerFactory;
+	}
+	
+    @Bean
+    public PasswordEncoder passwordEncoder(){
+        return new BCryptPasswordEncoder();
+    }
+
+	@Autowired
+	DataSource dataSource;
+	
+    @Bean
+	public JdbcUserDetailsManager jdbcUserDetailsManager() throws Exception {
+		JdbcUserDetailsManager jdbcUserDetailsManager = new JdbcUserDetailsManager();
+		jdbcUserDetailsManager.setDataSource(dataSource);
+		return jdbcUserDetailsManager;
 	}
 	
 	/*
